api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic[email]==2.5.3
python-dotenv==1.0.0
